requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pydantic==2.5.0
redis==5.0.1